[refactor] rename from index to student
